Extract classes and functions from activate
