+++ src/constants.ts
@@ -27,6 +27,7 @@ interface Neuro {
      */
     cancelled: boolean;
     outputChannel: vscode.OutputChannel | null;
+    chatHistoryChannel: vscode.OutputChannel | null;
     /** Whether Neuro has asked for a cookie. */
     waitingForCookie: boolean;
     tasks: NeuroTask[];
@@ -53,6 +54,7 @@ export const NEURO: Neuro = {
     waiting: false,
     cancelled: false,
     outputChannel: null,
+    chatHistoryChannel: null,
     waitingForCookie: false,
     tasks: [],
     currentTaskExecution: null,
